Fix test after rebase
