+++ python/ray/serve/_private/common.py
@@ -876,24 +876,6 @@ class CreatePlacementGroupRequest:
     fallback_strategy: Optional[List[Dict[str, Any]]] = None
 
 
-@PublicAPI(stability="alpha")
-@dataclass
-class GangContext:
-    """Context information for a replica that is part of a gang."""
-
-    gang_id: str
-    """Unique identifier for this gang."""
-
-    rank: int
-    """This replica's rank within the gang (0-indexed)."""
-
-    world_size: int
-    """Total number of replicas in this gang."""
-
-    member_replica_ids: List[str]
-    """List of replica IDs in this gang, ordered by rank."""
-
-
 @dataclass
 class GangPlacementGroupRequest:
     """Request to reserve gang placement groups for a deployment."""
