+++ python/ray/serve/_private/replica.py
@@ -48,7 +48,6 @@ from ray.serve import metrics
 from ray.serve._private.common import (
     RUNNING_REQUESTS_KEY,
     DeploymentID,
-    GangContext,
     ReplicaID,
     ReplicaMetricReport,
     ReplicaQueueLengthInfo,
